bioacoustics_inference_gui/requirements.txt
soundfile==0.12.1
scipy==1.10.1
Pillow==9.5.0
numpy==1.23.5
librosa==0.8.1
pandas==2.0.0
matplotlib==3.7.1
yattag==1.15.1
tensorflow==2.13.0rc1